Somewhat major overhaul to code structure.
Instead of all my "views" being UIViewControllers, they are now simply UIViews.
This makes sense from a hierarchy perspective and I probably should have realized it to begin with.
Luckily, it wasn't that hard to redesign.

Now, there is ONE UIViewController, that will manage how the views are placed.  My GPSView
still has the ability to push the ViewController containing a map onto the stack.